refactor(watchfs): move watchfs into fsi
watchfs should be completely owned by the fsi package / an fsi subsystem
in the future. This at least moves the package into the right directory
