+++ fsi/watchfs/watchfs_test.go
